Add new lane to a board
Added a '+' button to the top-right corner of the board that enables the
creation of a new lane on the board. The lane is created empty with the
name given by the user.

The add lane process calls the Board service to add a new lane to the
board, but it's still a dummy method that won't go to a server yet. This
needs to be done later when adding server-side behaviour to the software.

Fixes #6.